requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Pillow==10.0.0
pytesseract==0.3.10
firebase-admin==6.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
cryptography==41.0.3
reportlab==4.0.4
Flask-Babel==3.1.0
gunicorn==21.2.0
twilio==8.10.0 